Clean up storybook organization
