+++ src/stories/Home.stories.js
@@ -1,8 +1,6 @@
 import React from 'react'
 import Home from '../components/HomePage'
 import Help from '../components/Help/HelpPage'
-import Profile from '../components/Profile/ProfilePage'
-import ContactInfo from '../components/Profile/ContactInfo'
 
 export default {
   title: 'Home'
@@ -14,12 +12,4 @@ export const HomePage = () => (
 
 export const HelpPage = () => (
   <Help />
-)
-
-export const ProfilePage = () => (
-  <Profile />
-)
-
-export const ProfileContactInfo = () => (
-  <ContactInfo />
 )
