--- src/stories/Onboarding.stories.js
@@ -1,80 +1,0 @@
-import React from 'react'
-import Onboarding01 from '../components/Onboarding/01'
-import Onboarding02 from '../components/Onboarding/02'
-import Onboarding03 from '../components/Onboarding/03'
-import Onboarding04 from '../components/Onboarding/04'
-import Onboarding05 from '../components/Onboarding/05'
-import Onboarding06 from '../components/Onboarding/06'
-import Onboarding07 from '../components/Onboarding/07'
-import Onboarding08 from '../components/Onboarding/08'
-import Onboarding09 from '../components/Onboarding/09'
-import Onboarding10 from '../components/Onboarding/10'
-import Onboarding11 from '../components/Onboarding/11'
-import Onboarding12 from '../components/Onboarding/12'
-import Onboarding13 from '../components/Onboarding/13'
-import Onboarding14 from '../components/Onboarding/14'
-import PendingApprovalPage from '../components/PendingApprovalPage'
-
-export default {
-  title: 'Onboarding'
-}
-
-export const Onboarding01Page = () => (
-  <Onboarding01 />
-)
-
-export const Onboarding02Page = () => (
-  <Onboarding02 />
-)
-
-export const Onboarding03Page = () => (
-  <Onboarding03 />
-)
-
-export const Onboarding04Page = () => (
-  <Onboarding04 />
-)
-
-export const Onboarding05Page = () => (
-  <Onboarding05 />
-)
-
-export const Onboarding06Page = () => (
-  <Onboarding06 />
-)
-
-export const Onboarding07Page = () => (
-  <Onboarding07 />
-)
-
-export const Onboarding08Page = () => (
-  <Onboarding08 />
-)
-
-export const Onboarding09Page = () => (
-  <Onboarding09 />
-)
-
-export const Onboarding10Page = () => (
-  <Onboarding10 />
-)
-
-export const Onboarding11Page = () => (
-  <Onboarding11 />
-)
-
-export const Onboarding12Page = () => (
-  <Onboarding12 />
-)
-
-export const Onboarding13Page = () => (
-  <Onboarding13 />
-)
-
-export const Onboarding14Page = () => (
-  <Onboarding14 />
-)
-
-export const PendingApproval = () => (
-  <PendingApprovalPage />
-)
