+++ src/stories/Signup.stories.js
@@ -2,10 +2,7 @@ import React from 'react'
 import { LogIn } from '../components/Login'
 import { BecomeAmbassadorPage } from '../components/SignUp/BecomeAmbassadorPage'
 import { SignUpPage } from '../components/SignUp/SignUpPage'
-import { PersonalInfoPage } from '../components/SignUp/PersonalInfoPage'
-import { AddressPage } from '../components/SignUp/AddressPage'
-import { ContactPage } from '../components/SignUp/ContactPage'
-import { ContactInfoPage } from '../components/SignUp/ContactInfoPage'
+import PendingApprovalPage from '../components/PendingApprovalPage'
 
 export default {
   title: 'SignUp'
@@ -23,18 +20,6 @@ export const SignUp = () => (
   <SignUpPage />
 )
 
-export const PersonalInfo = () => (
-  <PersonalInfoPage />
-)
-
-export const Address = () => (
-  <AddressPage />
-)
-
-export const Contact = () => (
-  <ContactPage />
-)
-
-export const ContactInfo = () => (
-  <ContactInfoPage />
+export const PendingApproval = () => (
+  <PendingApprovalPage />
 )
